POPUG-014 Refactoring: separated the concern for creating a billing operations log entry into BillingOperationLogService.
