+++ employee-billing/src/main/java/org/uber/popug/employee/billing/service/BillingOperationLogService.java
@@ -1,0 +1,10 @@
+package org.uber.popug.employee.billing.service;
+
+import org.uber.popug.employee.billing.domain.aggregates.BillingOperationFullData;
+import org.uber.popug.employee.billing.domain.aggregates.TaskWithAssignee;
+
+public interface BillingOperationLogService {
+
+    BillingOperationFullData createBillingOperationLogEntry(TaskWithAssignee taskWithAssignee);
+
+}
